LPD-6611 Keep js organized in folders based on client extension type.
