+++ modules/apps/client-extension/client-extension-web/src/main/resources/META-INF/resources/js/index.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {default as FrontendTokenDefinitionFilePicker} from './components/FrontendTokenDefinitionFilePicker';
+export {default as FrontendTokenDefinitionFilePicker} from './components/theme-css/FrontendTokenDefinitionFilePicker';
 export {default as ModalImportClientExtensionEntry} from './components/ModalImportClientExtensionEntry';
 export {default as remoteProtocol} from './remote_protocol/bridge';
 export {default as ScriptElementAttributesFormField} from './components/global-js/ScriptElementAttributesFormField';
